fix: addressed feedback from pull request
- moved request body validation logic into a new validator class
- removed validation check 'Content Object contains more than one entry,' this is not a problem for request body content
