+++ src/utilities/validators/example-group-validator.ts
@@ -23,16 +23,6 @@ class ExampleGroupValidator extends BaseValidator {
     Object.entries(examples).forEach(([name, example]) => {
       this.checkExample(name, example);
     });
-
-    const requestBody = this.operation.requestBody;
-    if (requestBody) {
-      const example = this.operation.exampleRequestBody;
-      const content = requestBody.content;
-      const [key] = Object.keys(content);
-      const schema = content[key].schema;
-      const path: string[] = ['requestBody', 'example'];
-      this.validateObjectAgainstSchema(example, schema, path);
-    }
   };
 
   private checkMissingRequiredParameters(): void {
