+++ src/utilities/validators/index.ts
@@ -1,3 +1,4 @@
 export { default as ExampleGroupValidator } from './example-group-validator';
 export { default as ParameterSchemaValidator } from './parameter-schema-validator';
+export { default as RequestBodyValidator } from './request-body-validator';
 export { default as ResponseValidator } from './response-validator';
