+++ src/utilities/validators/parameter-schema-validator.ts
@@ -1,4 +1,4 @@
-import { ParameterObject, RequestBodyObject } from 'swagger-client/schema';
+import { ParameterObject } from 'swagger-client/schema';
 import {
   parameterHasContent,
   parameterHasSchema,
@@ -8,7 +8,6 @@ import {
   InvalidParameterExample,
   InvalidParameterObject,
   MissingContentSchemaObject,
-  InvalidRequestBodyContent,
 } from '../../validation-messages/failures';
 import OASOperation from '../oas-operation';
 import BaseValidator from './base-validator';
@@ -22,14 +21,13 @@ class ParameterSchemaValidator extends BaseValidator {
   }
 
   performValidation = (): void => {
-    if (this.operation.parameters !== undefined) {
-      this.operation.parameters.forEach((parameter) => {
-        this.checkParameterObject(parameter, ['parameters', parameter.name]);
-      });
+    if (this.operation.parameters === undefined) {
+      return;
     }
-    if (this.operation.requestBody !== undefined) {
-      this.checkRequestBody(this.operation.requestBody);
-    }
+
+    this.operation.parameters.forEach((parameter) => {
+      this.checkParameterObject(parameter, ['parameters', parameter.name]);
+    });
   };
 
   private checkParameterObject(
@@ -64,27 +62,6 @@ class ParameterSchemaValidator extends BaseValidator {
     } else if (!parameterHasSchema(parameter)) {
       // ERROR: Parameter Object must contain one of the following: schema, or content.
       this.addFailure(new InvalidParameterObject(path));
-    }
-  }
-
-  private checkRequestBody(requestBody: RequestBodyObject): void {
-    const path: string[] = ['requestBody'];
-
-    // unlike a parameter, the request body only has content, not schema
-
-    // check for invalid content
-    const [contentObjectKey, ...invalidKeys] = Object.keys(requestBody.content);
-    if (invalidKeys.length > 0) {
-      // ERROR: Content Object contains more than one entry.
-      this.addFailure(new InvalidRequestBodyContent([...path, 'content']));
-    }
-
-    // check for content schema presence
-    if (!requestBody.content[contentObjectKey].schema) {
-      // ERROR: Content Object does not contain a Schema Object.
-      this.addFailure(
-        new MissingContentSchemaObject([...path, 'content', contentObjectKey]),
-      );
     }
   }
 }
